MB-20359 Convert phosphor to a shared library
The purpose of this patch is to ensure that multiple components running
within the same process are all referring to the same singleton instance
of phosphor::TraceLog. For example with memcached and ep-engine, as
ep-engine does not link against the memcached daemon it must
independantly link with phosphor which would not have the desired effect
with a static library.

Because Windows cannot support exported thread-local symbols this patch
also de-inlines the ChunkTenant selection code. It also moves the
definition of the thread-id cache variable to be inlined so that it does
not need to be exported from the shared library.

Change-Id: Ic9ecdea7c9b22a3684fed91b7f98a833d6fed1a4
Reviewed-on: http://review.couchbase.org/66360
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>