MB-20428 Continue tracing event after resetting chunk
Prior to this patch the TraceLog would bail out of logging
an event if it had to reset the chunk (e.g. if tracing is stopped
and restarted while the current chunk is loaned out).

This patch introduces a small while loop that will retry acquiring
the lock after resetting.

Change-Id: I507a0587134c383bc238bee2d09d2690c2bf7269
Reviewed-on: http://review.couchbase.org/66541
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: buildbot <build@couchbase.com>