Refactor trace event macros and add tests for them
This patch refactors the trace event macros to consolidate
some of the code through a single set of macros (To minimise code
duplication amongst the macros when category filtering is
implemented within the macros).

It also adds a set of tests to run all of the public trace macros
and ensure they all work properly as one of them had an error
previously.

Change-Id: I6a95c4da427f484b00bf11e1a1ebf8778f310bfa
Reviewed-on: http://review.couchbase.org/66542
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: buildbot <build@couchbase.com>