Add PHOSPHOR_PTR macro for easy conversion to const void*
Phosphor cannot accept any pointer type, but it might on occasion
be useful to log the address of a given variable. In this sitation
it would make sense to convert the variable's pointer to void*
(Which phosphor accepts) but this can be a little messy/unclear.
Therefore this patch introduces a PHOSPHOR_PTR macro to make the
intent a little clearer.

Change-Id: I4b1d5466ee5354141b7d55c16d3290766afda5ec
Reviewed-on: http://review.couchbase.org/66650
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: buildbot <build@couchbase.com>