MB-20732 Refactor trace categories, names, arguments into a single object
This patch refactors the trace categories and names for each tracepoint
to be held in a single static object which allows for a reduction in
size of the TraceEvent.

It also moves the trace argument definitions to this tracepoint object
instead of being hardcoded in the serialisation code which will allow
for custom argument names to be added in a future patch.

Change-Id: Icf58325f02d10fd8d27c17b4a55844205715bddb
Reviewed-on: http://review.couchbase.org/67163
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: buildbot <build@couchbase.com>