MB-20732 Refactor trace categories, names, arguments into a single object
This patch refactors the trace categories and names for each tracepoint
to be held in a single static object which allows for a reduction in
size of the TraceEvent.

It also moves the trace argument definitions to this tracepoint object
instead of being hardcoded in the serialisation code which will allow
for custom argument names to be added in a future patch.

This readds a previous patch (1c64b10cd4c68d104c6aa5f5d3ec025b9fad93cf)
which was reverted as it wasn't quite threadsafe.

Change-Id: I23c24bdacd950884e599e9e468ba5b2889e49600
Reviewed-on: http://review.couchbase.org/67203
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: buildbot <build@couchbase.com>