Ensure memory is allocated & freed using same runtime
On Windows is it not safe in the general case to allocate memory in
one binary and deallocate from a different binary - for example new an
object in the main executable and delete it in a DLL. This is because
each executable can link to a different malloc implementation.

Typically this occurs if a different version of the CRT is used, but
in the case of phosphor's use in Couchbase Server this occurs when
allocation are made from memcached.exe (which uses cbmalloc as it
links libplatform) and deallocations are done from phosphor.dll (which
uses the system malloc).

The effect of this is hard-to-track down memory errors, or (if you're
luckier) crashes on free.

To solve this we need to ensure that any objects which are passed over
the phosphor DLL interface have matching allocation / free from one
side or the other. The following existing patterns breaks this rule,
and need to be fixed:

1. Constructors and destructors for objects which own memory must both
   be inline or both be non-inline, not mixed. This applies to both
   explicitly-declared con/destructors, and to compiler-generated
   defaults (which are implicitly inline).  The problem is that the
   inlined construtor/destuctor will allocate memory using the caller
   executable's malloc/free, whereas the non-inline will use
   phosphor's.  See: TraceConfig, JSONExport.

2. Non-const STL objects should not be returned from phosphor to the
   executable. This is similar to (1) - the STL object would have been
   allocated in phosphor, but may be freed (or resized) in the main
   executable. Solution is to either make const (so exe cannot free
   them), or create a wrapper class which we can ensure allocation and
   deletion occurs in the same place.  See: phosphor::String.

2b. Essentially the same issue as (2) - when passing callback objects
   into phosphor ensure that phosphor doesn't delete them as it didn't
   allocate the memory. This was an issue using std::function for
   callbacks, as the caller creates an object of this type, and then
   phosphor deletes it).  Solution is to instead use an abstract base
   class for the Callback, and have the concrete subclasses declares
   all constructors/destructors non-inline. To allow multiple entities
   to refer to a callback (without copying it), use shared_ptr to own
   it.  See: TracingStoppedCallback

Additionally, to defend against such issues in future we link the phosphor
unit tests with libplatform (and hence cbmalloc), so the unit tests run in
this mixed-mode environment.

Change-Id: I37ffcc8c1b0ffc06106bf21a11479f67df2ff0b1
Reviewed-on: http://review.couchbase.org/68398
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Reviewed-by: Jim Walker <jim@couchbase.com>