TRACE_LOCKGUARD: Provide the ability to event trace locks
Add a new macro which takes a std::mutex or similar (it needs to
provide a lock and unlock function), a category and a name. The
category and name match those used in the existing trace macros
e.g. TRACE_EVENT0.

The macro provides a scoped lock (i.e. the inputted lock is unlocked
on leaving the current scope).  In addition, the macro generates 4
events, that are used to record the duration of acquiring the lock,
and the duration the lock is held.

The acquiring phase is marked by a start and end event which has the
name <input name> + ".acquire".  The held phase is marked by a start
and end event which has the name <input name> + ".held".

A test for the new macro is also provided.

Change-Id: Ifde30a17cb4f7c4bc9d94b8ac46d82edee17c901
Reviewed-on: http://review.couchbase.org/69898
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>