Use #CPU-based thread count for benchmarks
A number of the benchmarks used a fixed (32) upper bound for the
number of threads to create. While this useful for cross-machine
comparisons, when running on machines whose CPU count is much less
than 32 (and/or virtualized hardware), running 32 threads can have a
significant detremental effect on the benchmark runtimes.

For example, running 32 threads on a 4-core Xen VM (ubuntu12-cv-05)
the following numbers are seen for chunk_replacement_benchmark:

    Run on (4 X 2497.18 MHz CPU s)
    2016-11-17 01:07:12
    Benchmark                              Time           CPU Iterations
    --------------------------------------------------------------------
    RegisterTenants/threads:1           1818 ns       1810 ns     388889
    RegisterTenants/threads:2           3336 ns       6638 ns     106060
    RegisterTenants/threads:4           2522 ns       9034 ns      82352
    RegisterTenants/threads:8          58773 ns     212704 ns      24560
    RegisterTenants/threads:16        321890 ns    1152500 ns       1600
    RegisterTenants/threads:32     224536907 ns  675466250 ns       3200

Compare these to a 24 core physical machine (same underlying hardware):

    Run on (24 X 2601 MHz CPU s)
    2016-11-17 01:08:37
    Benchmark                              Time           CPU Iterations
    --------------------------------------------------------------------
    RegisterTenants/threads:1           2449 ns       2447 ns     248977
    RegisterTenants/threads:2           3939 ns       7005 ns     157626
    RegisterTenants/threads:4           2351 ns       9031 ns      70964
    RegisterTenants/threads:8           2020 ns      11125 ns      60800
    RegisterTenants/threads:16          2143 ns      17306 ns      52800
    RegisterTenants/threads:32          2319 ns      16904 ns      36256

The times are comparable up to 4 threads (num CPUs for VM), but
degrade rapidly - the 32 thread test takes over 2milliseonds for an
iteration!

To allow benchmarks to be run on a wide range of hardware, remove the
fixed upport bound of 32, and instead use an upper bound of 2x the
number of CPUs.

To permit comparison between different sized machines, the upper bound
calculation can be overridden by setting the env var
PHOSPHOR_BENCH_NUM_THREADS - for example setting
PHOSPHOR_BENCH_NUM_THREADS=32 will give the same behaviour as before
this patch.

Change-Id: I793dff1473a26bd76546f6c5fab9f6965ef1c6cb
Reviewed-on: http://review.couchbase.org/70011
Reviewed-by: Daniel Owen <owend@couchbase.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Will Gardner <willg@rdner.io>