MB-21726: Set environment before running tests
If the PHOSPHOR_TRACING_START environment variable is set the
export_test, trace_log_test and memory_usage_test fail.

This patch fixes the problem by ensuring that the environment variable
is cleared before running any of the module tests or library tests.

To specify the enviroment it is best to have our own implementation of
main.  Therefore the use of gtest_main is removed for these tests.

In addition the Win32 implementation of setenv is moved into the file
containing main function (test_main.cc) allowing the removal of a
duplicate copy of the function.

Change-Id: I4e36d356ec02b4fc4517d62e4b32363be8ec296c
Reviewed-on: http://review.couchbase.org/70095
Reviewed-by: Will Gardner <willg@rdner.io>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: buildbot <build@couchbase.com>