Don't wait to stop tracing when the buffer is full
Previously threads would wait to acquire the TraceLog mutex in
order to stop tracing when the current trace buffer becomes full.

This can actually be quite expensive if the lock is being used for
anything other than stopping tracing that takes any decent length
of time as it will result in logEvent calls blocking on all
threads with full chunks until the lock is released.

Further it's a minor waste of time if another thread is already
in the process of shutting down tracing as it will end up
acquiring the lock only to discover that tracing has already been
stopped.

Change-Id: I27a771ea450316a45613b5379d139f6fc6a82856
Reviewed-on: http://review.couchbase.org/71401
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>