Implement the registry benchmark using condition variables
The benchmark was previously becoming deadlocked when run under
ThreadSanitizer on the CV slaves with more than 2 threads.

The reason behind this is unknown but forcing more of the logic
to be done under mutual exclusion via use of a condition variable
instead of an atomic and a busy loop seems to resolve this.

Change-Id: Ie48cda8dab11851c23a8866dd36f99ec7f576eed
Reviewed-on: http://review.couchbase.org/79653
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>