Remove shared chunk tenant support
Shared chunk tenants were originally added as a workaround for
ForestDB being used by Go components that would create very
shortlived threads.

With the likelyhood of phosphor now being used in ForestDB being
slim this patch removes the support for that.

As a side effect of doing this we now register a thread when it
tries to replace a chunk if it has not previously registered. There
is a (slight) potential for a SEGV to occur if a thread were to
subsequently terminate without deregistering and if tracing then
stopped as the TraceLog would try to evict a ChunkTenant allocated
in freed thread local storage.

This should not be an issue in KV-Engine as all threads (except
the main thread) are guaranteed to have registered and the
main thread stops tracing before exiting.

This will be resolved in a follow-up patch which will adjust the
way thread registration works to have ChunkTenant adopt RAII
(and instigate a temporary workaround for MacOS until a new
enough version of XCode is used to allow C++11 `thread_local`).

Change-Id: Ib1f53598dfadcdcb7842863505045ab63cd37724
Reviewed-on: http://review.couchbase.org/82500
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>