Introduce Complete event type
Add a new event type which logically combines a pair of SyncStart and
SyncEnd events into a single 'Complete' event.

In a trace that most of the events are duration events, using complete
events to replace the duration events can reduce the size of the trace
to about half.

This patch adds Complete events to the core of Phosphor; exposing them
via new TRACE_EVENT_COMPLETE* macros.

Change-Id: Ib445ea37d58ed5fb5f6b690c9c31d19337a09c56
Reviewed-on: http://review.couchbase.org/93877
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>