Allow for updating a TraceConfig from a config string
This is a minor refactor which separates the parsing and updating
logic from `TraceConfig::fromString` into its own function in
`TraceConfig::updateFromString` which allows a pre-existing
TraceConfig object to be updated from a config string.

This may be useful in a situation where the same config is to be
used with only minor adjustments (e.g. with only log size changed
but the same categories used).

Change-Id: Ifa3e923a6ae0a610e2c63f7a0409fe8df7be8c0d
Reviewed-on: http://review.couchbase.org/97794
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>