Store event and argument types in the TPI instead of TraceEvent
Event type and argument types _should_ be static for a given
tracepoint (since C++ is a statically typed language..) so this
reflects this in a space optimisation by moving those type
enums from the TraceEvent object to the tracepoint_info object.

This brings the nominal size of a TraceEvent from 47 bytes to
44 bytes (although this is still effectively 48 bytes due to
alignment).

Change-Id: If537a30da9b0f3e459c2db4aa133682cec31eee4
Reviewed-on: http://review.couchbase.org/97830
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>