Move thread id from the TraceEvent to the TraceChunk
Following 6cfacec the support for shared chunk tenants was removed
from phosphor. Since a chunk can no longer be used by multiple
threads the thread id for a given event can be determined by the
chunk it is in. This patch makes a space optimisation by moving
the thread id to the chunk instead of the event.

This reduces the nominal size of a TraceEvent object from 44 bytes
to 40 bytes (due to alignment this is actually a reduction from
48 bytes).

Change-Id: I664bae3970639023b8dad36aadfa6d549588fb8f
Reviewed-on: http://review.couchbase.org/97831
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>