De-inline the `logEvent` functions
This de-inlines/de-templates the TraceLog::logEvent
functions which drastically reduces the amount of
code generated in each compilation unit which uses
a phosphor tracepoint.

This also involved adding an explicit constructor
to TraceArgument for pointer types as the types
no longer get decayed by being passing through
the logEvent function before the constructor.

Change-Id: I58bce435fbfe76cda71671752efb41e55879d82c
Reviewed-on: http://review.couchbase.org/97832
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>