Correctly set TraceConfig::buffer_factory from string
When using TraceConfig::fromString the buffer factory returned
would always be the fixed trace buffer factory regardless of the
buffer mode. While the buffer mode member variable was being
correctly set in TraceConfig::updateFromString the buffer_factory
was not being kept in sync.

Since the dependance between the buffer mode and buffer factory is
so important (and easy to miss), to resolve this a utility class
TraceConfig::BufferFactoryContainer is introduced (ie. make it harder to
introduce the bug in the first place rather than just fix it).

In order to verify this, the TraceBuffer::bufferMode virtual member
function was added to determine the 'mode' of a buffer. The reason this
method was chosen is because TraceConfig::getBufferFactory returns a
std::function which makes determining the underlying function pointer
problematic as it would require use of type_info which behaves _very_
poorly across compilation units.

Change-Id: I563ebe0105d108e61a0aa27dc48ddbc496dac033
Reviewed-on: http://review.couchbase.org/98173
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>