MB-32016: Add 'unsanitized' variant of phosphor library
If one or more sanitizers are enabled (AddressSanitizer,
ThreadSanitizer, ...) then define an additional variant of libphosphor
which has the sanitizers /disabled/.  This allows targets which
require phosphor, but which are incompatible with the sanitizers
(e.g. Erlang NIFs use platform which uses phosphor) to still be built
and run when we are building the rest of the Server with sanitizer
support.

Targets which suffer from this limitation can link against
'phosphor_unsanitized' instead of 'phosphor' to obtain a library which
doesn't use any of the sanitizers.

Change-Id: Ic758d76d5bb7969a840b615644139a277fd06a08
Reviewed-on: http://review.couchbase.org/103245
Reviewed-by: Will Gardner <willg@rdner.io>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>