MB-42441: Don't auto-register ChunkTenant threads
Phosphor requires that any thread which records phosphor trace events
registers (and deregisters) itself on thread start / stop. this is
necessary so when phosphor stops, it can locate all threads which
potentially have a TraceChunk and "evict" them from the chunk (given
the chunk is about to be deleted).

However, for convenience if threads don't explicitly register
themselves, then they will get automatically registered as part of
acquiring their first TraceChunk, and hence can log events (albeit with
an empty thread name).

This is fine if the implicitly registered thread still exists when
Phosphor is shut down, however if that thread has already terminated
then a dangling ptr to the thread's ChunkTenant is left in
TraceLog::registered_chunk_tenants, which can result in segfault or
hang on shutdown when evictAllThreads attempts to lock and access it.

There's suggestion in the phosphor code
(http://review.couchbase.org/c/phosphor/+/82500) that auto
de-registration should be possible with full support for C++11
thread_local, by creating a dtor for ChunkTenant.  However doing so is
non-trivial - I encountered either data races when ChunkTenant dtor
dereferences a new ChunkTenant::traceLog member variable without a
lock; _or_ a deadlock if I attempt to use the ChunkLog to guard
access.

Therefore address this by removing the auto-registration - the
rationale being it's better to not record trace events for unknown
threads than it is to crash/hang on shutdown.

Change-Id: Iacff43073a23123931fcce3f1b1e47c477b88e90
Reviewed-on: http://review.couchbase.org/c/phosphor/+/140040
Reviewed-by: Jim Walker <jim@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>