MB-45044: Fix gcc-10 '-Wstringop-truncation' warnings
With the upgrade from GCC 7 to GCC 10, we see new warning of the form:

    phosphor/inline_zstring.h:55:20: ‘char* strncpy(char*, const char*, size_t)’ output truncated copying 8 bytes from a string of length 13 [-Wstringop-truncation]
       55 |             strncpy(_s, s, max_length);
          |             ~~~~~~~^~~~~~~~~~~~~~~~~~~

This is a result of the stricter 'stringop-truncation' checks which
warn if strncpy input is truncated during copy. In this particular
instance we don't require the destination buffer to be null-terminated
(if at max length) so the warning is benign, however given it's a
useful warning in general so keep it enabled, but use memcpy / memset
instead to explicilty set the entire inline string.

Change-Id: I8e0f2dc3c11d67ab56ca666d24e1f561b9e270a5
Reviewed-on: http://review.couchbase.org/c/phosphor/+/150261
Tested-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>