Improve precision of JSON exported timestamp field
Google tracing expects the timestamp of events to be represented as
microseconds, which is a loss in precision compared to the nanoseconds
we record. This can result in the Google Trace render getting mixed up
which events were parents of what, as we simply discard the fractional
part of the timestamp.

Avoid this by keeping the remainder of the division, and including as
fractional microseconds (which Chrome's trace viewer accepts).

Change-Id: I76687c8388ef4efd76073f4510c721d5490f5a39
Reviewed-on: https://review.couchbase.org/c/phosphor/+/185025
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>