MB-58833: Fix gcc -Wshadow=compatible-local warnings
We have recently hit a nasty issue caused by variable shadowing and the
consequent usage of the wrong variable reference (MB-58463).

The -Wshadow flag would give us a compilation warning and reduce the
chance of making that kind of error.

Here we fix phosphor before enabling the flag in tlm.

Change-Id: I62b30da2b2311e4f0d32d6eca4c3f00a1c85bd0f
Reviewed-on: https://review.couchbase.org/c/phosphor/+/197947
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>