MB-66284: Flip order for TraceConfig and TraceLogConfig
TraceLogConfig contains a std::unique_ptr of TraceConfig but
is defined before the TraceConfig. This caused a problem when
flipping to C++23 where it failed due to not knowing the
size of the object or its destructor.

Given that TraceConfig don't need anything from TraceLogConfig
it can be defined before TraceLogConfig and work around the
problem

Change-Id: Id6eac0381ea8ffa706f3578a55050941fb1584c8
Reviewed-on: https://review.couchbase.org/c/phosphor/+/226364
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@couchbase.com>